fix: Add support for per-task resource requests + limits in K8s executor

+++ cmd/util/flags.go
@@ -109,7 +109,7 @@ func workerFlags(flagConf *config.Config) *pflag.FlagSet {
 func nodeFlags(flagConf *config.Config) *pflag.FlagSet {
 	f := pflag.NewFlagSet("", pflag.ContinueOnError)
 
-	f.Uint32Var(&flagConf.Node.Resources.Cpus, "Node.Resources.Cpus", flagConf.Node.Resources.Cpus, "Cpus available to Node")
+	f.Float64Var(&flagConf.Node.Resources.Cpus, "Node.Resources.Cpus", flagConf.Node.Resources.Cpus, "Cpus available to Node")
 	f.Float64Var(&flagConf.Node.Resources.RamGb, "Node.Resources.RamGb", flagConf.Node.Resources.RamGb, "Ram (GB) available to Node")
 	f.Float64Var(&flagConf.Node.Resources.DiskGb, "Node.Resources.DiskGb", flagConf.Node.Resources.DiskGb, "Free disk (GB) available to Node")
 	f.Var(&TimeoutConfigValue{&flagConf.Node.Timeout}, "Node.Timeout", "Node timeout in seconds")
